% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{etl}
\alias{etl}
\title{Import raw data to corresponding object}
\usage{
etl(
  which = c("grants", "publications", "presentations"),
  uri,
  token,
  use_cache = TRUE,
  verbose = TRUE,
  file = NULL
)
}
\arguments{
\item{which}{The object type to load (grants, publications, presentations)}

\item{uri}{REDCap URI to import from}

\item{token}{REDCap URI Authentication token for importing}

\item{use_cache}{If previously run, the cached object can be used instead of reloading}

\item{verbose}{Should the process be verbose so the steps can be viewed}

\item{file}{Output file to store results (useful for caching).}
}
\value{
An object (grants,publications, presentations) which is mostly
a data frame with an attached data dictionary.
}
\description{
Import REDcap data into corresponding object and transform
variables into usable reporting format.
}
\details{
This function is likely the main driver of functionality within
the library. Specifically, this function will import data from redcap
(possibly using cached data if requested) then transform variables into
useful reporting variables suitable for the library \code{pgreportr}.

There are three different input types (grants, publications and
presentations) that this function will handle.
}
