% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_investigators}
\alias{extract_investigators}
\title{Create a table specific for grant investigator information.}
\usage{
extract_investigators(redcap_data)
}
\arguments{
\item{data}{Raw source data}
}
\value{
A tibble corresponding to investigators.
}
\description{
Grant investigator information is repeated rows from the raw data
}
\details{
The data is formatted for data entry, not reporting, so there are
many different columns representing the same type of information.
}
