% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fapply.R
\name{fapply}
\alias{fapply}
\title{Apply a list of functions to a data frame}
\usage{
fapply(x, flist, verbose = TRUE)
}
\arguments{
\item{x}{A data frame}

\item{flist}{A list of functions}
}
\value{
A modified data frame resulting from applying the functions in \code{f}.
}
\description{
Apply a list of functions to a data frame
}
\examples{
fapply(iris[,1:4], list(function(x){x+1}, function(x){x-1}))
# This should result in iris[,1:4] as result (less rounding errors).
}
