% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{get_cache_file}
\alias{get_cache_file}
\title{Return name of REDCap cache file}
\usage{
get_cache_file(uri, token)
}
\arguments{
\item{uri}{A URI that was originally used to retrieve results.}

\item{token}{A token corresponding to the project in URI.}
}
\value{
A full pathname to a cache file for the URI
}
\description{
Return name of REDCap cache file
}
