% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkbox.R
\name{get_field_choices}
\alias{get_field_choices}
\title{Return the choices for a field from a grant/pub table}
\usage{
get_field_choices(x, field)
}
\arguments{
\item{x}{A grant/pub table}

\item{field}{The field name to return values for}
}
\description{
Return the choices for a field from a grant/pub table
}
