% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import pgreportr object}
\usage{
import(
  uri,
  token,
  which = c("grants", "publications", "presentations"),
  use_cache = TRUE,
  verbose = TRUE
)
}
\description{
General-purpose routines for importing data
from REDCap for pgreportr library
}
\details{
The \code{pgreportr} library manages two different object
types: grants and publications. The overall process is to import
this data directly from REDCap (using the \code{\link[REDCapR:redcap_read]{REDCapR::redcap_read()}})
function. And then perform a number of data manipulations on the
imported data to make it suitable for later reporting.

Therefore each individual import: \code{\link[=import_grants]{import_grants()}}
and \code{\link[=import_publications]{import_publications()}} will call \code{\link[=read_redcap_data]{read_redcap_data()}}
and then do followup processing. See the help items for
\code{\link[=import_grants]{import_grants()}} and \code{\link[=import_publications]{import_publications()}} for specifics.

Where possible, common functionality, code and terminology is used
between objects. For instance, the concept of a program grant
year, creators (whether grant investigators or publication authors).
}
