% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_grants}
\alias{import_grants}
\title{Import REDCap export data into R for reporting.}
\usage{
import_grants(uri, token, use_cache = TRUE, verbose = TRUE)
}
\arguments{
\item{use_cache}{(Default: TRUE) Should a cached copy of the REDCap data be used (if available)?}

\item{verbose}{(Default: FALSE) Should messages be logged to the console?}
}
\value{
A tibble with grant data available.
}
\description{
The REDCap data can be exported in R format. This consists of an
R script and corresponding CSV to read in. Labels are defined,
factors are created, etc.
}
\details{
There are a few caveats to using this type of data which this function
overcomes. The result is a similar data structure, but cleaned up for use
in this library.

Note that if you provide the \code{use_cache} option, the data will be loaded
from local cache (if available).
}
