% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{import_redcap_data}
\alias{import_redcap_data}
\title{Read REDCap data with caching}
\usage{
import_redcap_data(uri, token, use_cache = TRUE, verbose = TRUE)
}
\arguments{
\item{uri}{URI of the REDCap server project}

\item{token}{Token for access authorization}

\item{use_cache}{Should a cached copy of the data be used instead of
querying the URI?}

\item{verbose}{(default: TRUE) Should messages be printed to console during operation.}
}
\value{
A list of raw, labelled and data dictionary (see \code{\link[=read_redcap_data_online]{read_redcap_data_online()}})
}
\description{
Load REDCap data from server with built-in support
for caching of results
}
\details{
This function is the main entry-point for loading REDCap pgreportr data
into R. It will managing caching of the results, so that subsequent requests
do not have to hit the server.

The \code{\link[=read_redcap_data_online]{read_redcap_data_online()}} function includes details on the process
and the return result.
}
