% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{import_redcap_data_online}
\alias{import_redcap_data_online}
\title{Retrieve REDCap data from the server}
\usage{
import_redcap_data_online(uri, token, verbose = TRUE)
}
\arguments{
\item{uri}{URI of the redcap data source (general api path)}

\item{token}{An authentication token to use when retrieving the data. Note that
the token indicates the project/forms to download.}

\item{verbose}{(default: TRUE) Should messages be printed to console during operation.}
}
\value{
A list consisting of \code{raw}, \code{label} and \code{dictionary} tables for the given
URI reference.
}
\description{
Using the REDCapR library, import the raw and labelled versions of
a project; and the data dictionary.
}
\details{
The overall process is to import
data directly from REDCap (using the \code{\link[REDCapR:redcap_read]{REDCapR::redcap_read()}}
function). The raw and labelled data are both read so that various mappings can
be created. Additionally, the metadata is imported as a data dictionary.

The result of this function is a list:
\itemize{
\item raw (unlabelled data)
\item label (labelled data)
\item dictionary (corresponding data dictionary)
}
}
\note{
The input token defines the specific project forms to be imported. This can
be generated from within REDCap.
}
