% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_grant_table}
\alias{is_grant_table}
\title{Test if data frame is a grant table}
\usage{
is_grant_table(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
Logical indicating if data frame is a grant table
}
\description{
Indicate if the passed data frame is a
grant table.
}
\details{
A grant table is a data frame (tibble) that has
certain columns available for processing. For convenience,
we attach an attribute to the table for indicating this
fact.
}
