% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_grant.R
\name{new_program_grant}
\alias{new_program_grant}
\title{Create a program grant object from a configuration list}
\usage{
new_program_grant(pg)
}
\arguments{
\item{pg}{}
}
\description{
Create a program grant object from a configuration list
}
