% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_grant.R
\name{pg_final_year}
\alias{pg_final_year}
\title{Return the last year of the Partnership Grant}
\usage{
pg_final_year(pg = program_grant())
}
\value{
An string representing the last (max) grant year.
}
\description{
Return the last year of the Partnership Grant
}
