% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_grant.R
\name{pg_project_period}
\alias{pg_project_period}
\title{Project period information from program grant}
\usage{
pg_project_period(d = NULL, year = NULL, pg = program_grant())
}
\arguments{
\item{d}{An optional date to report program grant project period}

\item{year}{An optional named year for the program project}

\item{pg}{A program grant (optional)}
}
\value{
A table consisting of information about the requested project period.
}
\description{
Retrieve program grant information for the different
periods of activity as a table.
}
\details{
The program project has activity across a number of years. This
function can provide the information about these years. Note that the time
periods are not necessarily aligned with calendar, fiscal or other boundaries.
}
