% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_grant.R
\name{pg_project_year}
\alias{pg_project_year}
\title{Project year from program grant}
\usage{
pg_project_year(d = NULL, year = NULL, pg = program_grant(), as_factor = FALSE)
}
\arguments{
\item{pg}{}
}
\description{
Project year from program grant
}
