% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Selecting variables for object}
\description{
Tools for selecting column subsets from raw data
}
\details{
There are a number of columns available for processing in the grants,
publications and presentations. Some are specific to the grant, some
are specific to the investigator (individual). These routines help
select the various subsets. Specifically, they are implemented by
'creator': investigator, author or presenter. There is an option to
negate the selection, allowing for both selecting of the grant/pub or
the creator(s).
}
