% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_author_columns}
\alias{select_author_columns}
\title{Remove author fields from publication tibble}
\usage{
select_author_columns(x, negate = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A tibble containing publications and authors.}

\item{negate}{(FALSE) Should author columns be extracted (FALSE) or removed (TRUE)}

\item{verbose}{Report the number of columns removed to console}
}
\value{
A data frame with the Author-related columns selected (or removed).
}
\description{
The publication table includes rows (and columns)
representing authors (one row per author). This function keeps
the rows and columns belonging to authors specifically.
}
