% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_investigator_columns}
\alias{select_investigator_columns}
\title{Select Investigator-related columns from data frame}
\usage{
select_investigator_columns(x, negate = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A data frame imported from redcap form}

\item{negate}{(FALSE) Should investigator columns be extracted (FALSE) or removed (TRUE)}

\item{verbose}{Report the number of columns removed to console}
}
\value{
A data frame with the Investigator-related columns selected (or removed).
}
\description{
Select Investigator-related columns from data frame
}
