% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_presenter_columns}
\alias{select_presenter_columns}
\title{Remove presenter fields from presentations tibble}
\usage{
select_presenter_columns(x, negate = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{A tibble containing presentations and presenters}

\item{negate}{(FALSE) Should presenter columns be extracted (FALSE) or removed (TRUE)}

\item{verbose}{Report the number of columns removed to console}
}
\value{
A data frame with the Presenter-related columns selected (or removed).
}
\description{
The presentation table includes rows (and columns)
representing presenters (one row per presenter). This function keeps
the rows and columns belonging to presenters specifically.
}
