% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_redcap_control_columns}
\alias{select_redcap_control_columns}
\title{Select REDCap-related columns}
\usage{
select_redcap_control_columns(x, negate = FALSE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
There are a few redcap control columns in the raw export
that may not be not needed (beyond initial processing).
}
