% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_config}
\alias{set_config}
\title{Internal function for processing config object}
\usage{
set_config(pg_config)
}
\arguments{
\item{pg_config}{}
}
\value{
The program grant object is returned (invisibly).
}
\description{
The config information read from \code{config.yml} must
be translated into usable library options. This function does
that transformation.
}
\details{
This is not likely a function useful outside of the
library itself. Briefly, a configuration is
read from a file (\code{config.yml}). This function will take that
very nested list of information and separate it into a program_grant
object and a set of library options (defaults).

The program_grant object is stored as \code{pgimportr.program_grant}
and the options are stored as \code{pgimportr.option} in the \code{\link[=options]{options()}}
variable.
}
