% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Dispatch transformation of raw data}
\usage{
transform(obj)
}
\arguments{
\item{x}{A raw REDCap data frame to transform}
}
\value{
A transformed object (grants, publications, presentations) after
appropriate ETL.
}
\description{
A series of ETL scripts applied to raw input data
}
\details{
The derivation of grants, publications and presentations occurs
via a series of transformation scripts. These functions perform this
ETL.

The overall process is:
\itemize{
\item Download REDCap data and data dictionary
\item Convert this data (as tibbles) into an object (grants, publications, presentations).
Note this is mostly to keep the data dictionary with the object, and to help in
data verification via typing.
\item Apply a series of ETL scripts to object in order to transform it into usable
reporting data.
}

These functions are the last step - transforming the raw input into usable fields.
Steps include things like converting checkboxes into structured variables, lists,
normalizing variable names and other pieces.

Code is included with the package as separate R files for each step of the
transformation (numbered so that the process occurs in order). That is, there
is a package \code{R.d/publications} directory with a number of scripts of the
form \code{05-do_this_operation.R}. Each contains a function that accepts a data
frame as input and returns a data frame, presumably somewhat transformed by
operations within the function.

The process is generic enough that additional \code{R.d/publications} entries
can be added/modified where needed without impacting these driver functions.
}
