% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_author_columns}
\alias{select_author_columns}
\title{Remove author fields from publication tibble}
\usage{
select_author_columns(x, negate = FALSE, verbose = FALSE)
}
\arguments{
\item{d}{A tibble containing publications and authors.}
}
\value{
A modified tibble with publication entries removed.
}
\description{
The publication table includes rows (and columns)
representing authors (one row per author). This function keeps
the rows and columns belonging to authors specifically.
}
